우선도를 구하는 방법에 치명적인 오류가 있었습니디ㅏ. BindViewHolder에서 우선도를 구해서 넣어주면  viewhodler가 만들어 지지 않으면 계산이 안되서 우선도가 그냥 초기값이 들어가 버립니다. 따라서 우선도 순의 정렬이 되지 않고, 엉망 진창의 순서가 됩니다. 이를 고치기 위해서, 따로 recylcerView에 우선도를 구하는 함수를 만들고, 외부에서 adpater.Calculate 으로 우선 우선도를 먼저 구한 다음에 들어가도록 했습니다
