+++ app/src/main/java/com/example/whattodo/search/SearchFragment.kt
@@ -42,36 +42,8 @@ class SearchFragment : Fragment() {
     private fun setupRecyclerView() {
         viewBinding.searchRecyclerView.layoutManager = LinearLayoutManager(context)
         searchRecyclerAdapter = MyAdapter(ToDo.previewData)
-        searchRecyclerAdapter.itemClickListener = object : MyAdapter.OnItemClickListener{
-            override fun OnItemClick(position: Int) {
-                searchRecyclerAdapter.setPriorityColor("2f22e0", "ca22e0", "db184f")
-                searchRecyclerAdapter.sortItemwithAscendingPriority()
-            }
-
-        }
-
-        searchRecyclerAdapter.calculatePriorityListener = object : MyAdapter.OnCalculatePriorityListener{
-            override fun calculatePriority(
-                _importance: Int,
-                _timeLeft: Long,
-                _time_taken: Float
-            ): Float {
-                val timeLeft = (_timeLeft / (60 * 60 * 1000)).toInt() // 남은 시간
-                var spareTime = timeLeft - _time_taken
-
-                if(timeLeft < 0)
-                {
-                    return -1.0f // 아예 기간이 지나면 음수를 반환함
-                }
-
-                if(spareTime < 0) // 만약 남은 시간 보다 소요 시간이 더 걸리면
-                {
-                    spareTime = 0.001f // 극단적으로 줄여서 우선도 상에서 매우 높은 비중을 가지게 해준다
-                }
-
-                return 1 / spareTime + _importance * 10
-            }
-        }
+        searchRecyclerAdapter.CalcItemsPrority()
+        searchRecyclerAdapter.sortItemwithDescendingPriority()
         viewBinding.searchRecyclerView.adapter = searchRecyclerAdapter
     }
