+++ app/src/main/res/drawable/recyclerview_style.xml
@@ -7,7 +7,7 @@
         />
 
     <corners
-        android:radius="7dp"
+        android:radius="12dp"
         />
 
 </shape>
