+++ app/src/main/res/layout/fragment_deadline.xml
@@ -33,7 +33,7 @@
             android:layout_height="0dp"
             android:layout_marginStart="34dp"
             android:layout_marginEnd="33dp"
-            android:layout_marginBottom="44dp"
+            android:layout_marginBottom="0dp"
             android:background="@drawable/recyclerview_style"
             android:padding="10dp"
             app:layout_constraintBottom_toBottomOf="parent"
